Refactor runtime assembly layering and policy separation
