+++ core/policy/__init__.py
@@ -1,0 +1,4 @@
+"""Domain policies governing payload transformation rules."""
+from .payload import classify_media, resolve_caption_policy
+
+__all__ = ["classify_media", "resolve_caption_policy"]
